Forms

See merge request Mianaan/phat_advisor!36

+++ public/add-favorite-restaurant.php
@@ -1,31 +1,28 @@
 <?php
-require '../vendor/autoload.php';
 
+require '../vendor/autoload.php';
+header("Access-Control-Allow-Origin: *");
 session_start();
 
-$userRepository = new \Repository\User();
 $restoRepository = new \Repository\Restaurant();
 
-$error = [];
-
-if (isset($_SESSION['id'])) {
+$error = "null";
+if ($_SERVER['REQUEST_METHOD'] !== 'GET'/* || !isset($_SESSION['id'])*/) {
+    $error = "internal error";
+    http_response_code(400);
+}else{
     $id_user = $_SESSION['id'];
     $id_restaurant = $_POST['id_resto'];
+    $is_favorite = $_POST['id_resto'];
 
-if($restoRepository->isAlreadyFavorite($id_user, $id_restaurant)){
-    $restoRepository->deleteFavoriteById($id_user, $id_restaurant);
-    $error = "deleted";
-}else{
+    if ($restoRepository->isAlreadyFavorite($id_user, $id_restaurant) && $is_favorite) {
+        $restoRepository->deleteFavoriteById($id_user, $id_restaurant);
+        $error = "deleted";
+    } else {
+        $restoRepository->addFavorite($id_user, $id_restaurant);
+        $error = "added";
+    }
 
-        if (!$restoRepository->addFavorite($id_user, $id_restaurant)) {
-            $error = "cant add in favorite";
-        }else{
-            $error = "added";
-        }
-
-} 
-}else{
-    http_response_code(400);
 }
 
 echo json_encode($error);
